+++ src/main/java/com/store/back/repositories/ProductRepository.java
@@ -2,7 +2,7 @@ package com.store.back.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.store.back.models.tablas.Product;
+import com.store.back.models.tables.Product;
 
 public interface ProductRepository extends JpaRepository<Product, Integer> {
