+++ src/main/java/com/store/back/services/ProductServiceImpl.java
@@ -8,7 +8,7 @@ import java.util.stream.Collectors;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.store.back.models.tablas.Product;
+import com.store.back.models.tables.Product;
 import com.store.back.models.DTO.ProductDto;
 import com.store.back.repositories.ProductRepository;
